requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pillow==11.0.0
PyYAML==6.0.2
rembg==2.0.60
requests==2.32.3
selenium==4.26.1
